Fix mistake in `find_equilibrium_position` where bounds were swapped.
Also scope the `position_cost` function within `find_equilibrium_position` since it only makes sense in that context.

I'm surprised that the results are identical before and after this commit:

[crates/buoyancy/src/lib.rs:260:11] results = Ok(
    (
        FtolReached,
        [
            -0.2500000104901638,
            -0.0454084412638515,
        ],
        6.796124273397224e-5,
    ),
)

which suggests to me that the optimizer is not sufficiently exploring the angular space. (Which is also why the found position isn't fully rotated to the correct floating position.)